fadeIn() and fadeOut() - Fix issue #180
The fadeOut() method my is own, while fadeIn() is a simplified version of: https://stackoverflow.com/a/20533102/4118605

There are more upvoted examples in SO, but this one seemed to me particularly apt, because it allowed (as the jQuery method) to set the requested timing in ms.

There is some inconsistency "behind the scenes" between the two, since only fadeOut is using CSS3 transitions. Alas, I couldn't come up quickly with a better solution.